requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
aiosqlite>=0.20,<1.0
pydantic-settings>=2.6,<3.0
httpx>=0.27,<1.0
Pillow>=10.4,<12.0
boto3>=1.36,<2.0
pytest>=8.3,<9.0
